Extract image generation into Cache\Stats_Image class
- Move image generation, caching, color/font resolution from the REST
  controller into a dedicated Cache\Stats_Image class.
- Controller is now a thin layer with two endpoints:
  /stats/image/{user_id}/{year} serves the image binary,
  /stats/image-url/{user_id}/{year} returns the resolved URL as JSON.
- Cache stores images in uploads/activitypub/stats/{user_id}/ using
  WP_Image_Editor for WebP optimization.
- Add activitypub_stats_image_url filter for CDN/Photon integration.
- Add activitypub_cache_stats_image_enabled filter to disable caching.
- Editor sidebar fetches the resolved URL via the image-url endpoint.
- ActivityPub attachment uses the direct cached file URL.
